requirements.txt
django-debug-toolbar==1.11
django-environ==0.4.5
django-extensions==2.1.7
django-redis==4.10.0
Django==2.2.1
djangorestframework==3.9.4
git+https://github.com/Supervisor/supervisor.git@master#egg=supervisor
git+https://gitlab.com/pycqa/flake8@master#egg=flake8
gunicorn==19.9.0
ipdb==0.12
psutil==5.6.2
psycopg2-binary==2.8.2
pyephem==3.7.6.0
pytest-django==3.4.8
redis==3.2.1
requests==2.22.0
timezonefinder==4.0.2
tornado==6.0.2
